Split cache lock tests for Memcached and Redis
The previous setup incorrectly skipped even the Redis tests when
the Memcached extension was not installed. By separating them, we
can also separate the logic for skipping these tests.
